Move loadBundle into GtfsBundle and rename it to loadDao.
