+++ application/src/main/java/org/opentripplanner/gtfs/graphbuilder/GtfsBundle.java
@@ -1,14 +1,39 @@
 package org.opentripplanner.gtfs.graphbuilder;
 
+import java.io.IOException;
 import java.io.InputStream;
+import java.util.Set;
 import javax.annotation.Nullable;
 import org.onebusaway.csv_entities.CsvInputSource;
+import org.onebusaway.gtfs.impl.GtfsRelationalDaoImpl;
+import org.onebusaway.gtfs.model.Area;
+import org.onebusaway.gtfs.model.FareLegRule;
+import org.onebusaway.gtfs.model.FareMedium;
+import org.onebusaway.gtfs.model.FareProduct;
+import org.onebusaway.gtfs.model.FareTransferRule;
+import org.onebusaway.gtfs.model.RiderCategory;
+import org.onebusaway.gtfs.model.RouteNetworkAssignment;
+import org.onebusaway.gtfs.model.StopAreaElement;
+import org.onebusaway.gtfs.serialization.GtfsReader;
+import org.onebusaway.gtfs.services.GtfsRelationalDao;
 import org.opentripplanner.datastore.api.CompositeDataSource;
+import org.opentripplanner.framework.application.OTPFeature;
 import org.opentripplanner.gtfs.config.GtfsFeedParameters;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
 public final class GtfsBundle {
+
+  private static final Set<Class<?>> FARES_V2_CLASSES = Set.of(
+    Area.class,
+    FareProduct.class,
+    FareLegRule.class,
+    FareMedium.class,
+    FareTransferRule.class,
+    RiderCategory.class,
+    RouteNetworkAssignment.class,
+    StopAreaElement.class
+  );
 
   private static final Logger LOG = LoggerFactory.getLogger(GtfsBundle.class);
 
@@ -81,5 +106,38 @@ public final class GtfsBundle {
 
   public String feedInfo() {
     return "GTFS bundle at " + dataSource.path() + " (" + getFeedId() + ")";
+  }
+
+  public GtfsRelationalDao loadDao() throws IOException {
+    var dao = new GtfsRelationalDaoImpl();
+    dao.setPackShapePoints(true);
+    LOG.info("reading {}", feedInfo());
+
+    GtfsReader reader = new GtfsReader();
+    reader.setInputSource(getCsvInputSource());
+    reader.setEntityStore(dao);
+    reader.setInternStrings(true);
+    reader.setDefaultAgencyId(getFeedId());
+
+    dao.open();
+    for (Class<?> entityClass : reader.getEntityClasses()) {
+      if (skipEntityClass(entityClass)) {
+        LOG.info("Skipping entity: {}", entityClass.getName());
+        continue;
+      }
+      LOG.info("Reading entity: {}", entityClass.getName());
+      reader.readEntities(entityClass);
+    }
+    dao.close();
+    return dao;
+  }
+
+  /**
+   * Since GTFS Fares V2 is a very new, constantly evolving standard there might be a lot of errors
+   * in the data. We only want to try to parse them when the feature flag is explicitly enabled as
+   * it can easily lead to graph build failures.
+   */
+  private boolean skipEntityClass(Class<?> entityClass) {
+    return OTPFeature.FaresV2.isOff() && FARES_V2_CLASSES.contains(entityClass);
   }
 }
